Credit fldef - These changes allow disc swapping.
This code was entirely made by fldef. https://github.com/fldef/pcsx2-fldef
Unfortunately he retired his account and disappeared a month ago with no signs that he will ever reappear and no way to contact him. The purpose of this isn't to steal code from him but to keep an active repository with it. I'm not creating a PR with the hope he will come back and make a PR himself.